requirements.txt
Package             Version
------------------- ---------
asgiref             3.3.1
atomicwrites        1.4.0
attrs               20.3.0
certifi             2020.11.8
chardet             3.0.4
colorama            0.4.4
Django              3.1.3
django-filter       2.4.0
djangorestframework 3.12.2
idna                2.10
importlib-metadata  2.0.0
iniconfig           1.1.1
Markdown            3.3.3
packaging           20.4
pip                 20.2.4
pluggy              0.13.1
py                  1.9.0
pyparsing           2.4.7
pytest              6.1.2
pytz                2020.4
requests            2.25.0
setuptools          50.3.2
six                 1.15.0
sqlparse            0.4.1
toml                0.10.2
urllib3             1.26.2
wheel               0.35.1
zipp                3.4.0
